gold-tier/requirements.txt
# Digital FTE - Gold Tier Requirements
# Core dependencies for AI Employee System

# HTTP and API clients
requests>=2.32.0
httpx>=0.27.0

# Facebook/Meta SDK
facebook-business>=21.0.0

# Twitter API
tweepy>=4.14.0

# Odoo Integration
xmlrpcclient2>=1.0.3
zeep>=4.2.1

# MCP Server Framework
mcp>=1.0.0

# File watching
watchdog>=4.0.0

# Google API (Gmail)
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Browser automation
playwright>=1.45.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# Date/time handling
python-dateutil>=2.8.2
pytz>=2024.1

# Logging and error handling
structlog>=24.1.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Text processing
markdown>=3.5.0
jinja2>=3.1.2

# Sentiment analysis (for social media monitoring)
textblob>=0.17.1

# Testing (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
